refactor: 리팩토링 후 예전 코드 삭제
