+++ src/test/java/nextstep/auth/authorization/TokenSecurityContextPersistenceInterceptorTest.java
@@ -5,23 +5,19 @@ import nextstep.auth.domain.Authentication;
 import nextstep.auth.infrastructure.JwtTokenProvider;
 import nextstep.auth.infrastructure.SecurityContextHolder;
 import nextstep.auth.ui.authorization.SecurityContextInterceptor;
-import nextstep.auth.ui.authorization.SessionSecurityContextPersistenceInterceptor2;
-import nextstep.auth.ui.authorization.TokenSecurityContextPersistenceInterceptor2;
+import nextstep.auth.ui.authorization.TokenSecurityContextPersistenceInterceptor;
 import nextstep.member.domain.LoginMember;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mock;
-import org.mockito.Mockito;
 import org.mockito.Spy;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.mock.web.MockHttpServletRequest;
 import org.springframework.mock.web.MockHttpServletResponse;
 
 import java.io.IOException;
-import java.util.Map;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
@@ -48,7 +44,7 @@ public class TokenSecurityContextPersistenceInterceptorTest {
 
     @BeforeEach
     void setUp() {
-        interceptor = new TokenSecurityContextPersistenceInterceptor2(jwtTokenProvider, objectMapper);
+        interceptor = new TokenSecurityContextPersistenceInterceptor(jwtTokenProvider, objectMapper);
         response = new MockHttpServletResponse();
         loginMember = new LoginMember(1L, EMAIL, PASSWORD, AGE);
